Service now allows you to use a default local or remote YARA rule set. The rules are pulled and compiled on the Docker container build.
The service worker also has the base code needed to allow for custom rules to be sent during tasking.

Additionally, this change switch yara.compile to yara.load to provide some performance boost. This was done becasue we are compiling the rules on the Docker build and only allowing compiles custom rules.