Heavy modifications.
Outsourced most of the actual installation process to subdirectories of the install folder.
This makes the install.sh much more readable.
Since the Dockerfile in the Docker directory actually belongs to the totem install process it has been moved to install/totem/totem.dockerfile.
Fixed the upstart scripts and hopefully the systemd scripts as well.

TODO: Test installation on a system running with systemd.

Signed-off-by: Maximilian Schott <ms@roottec.com>